move forum's privileges
