+++ lib/api-v2/db-api/comments/index.js
@@ -1,5 +1,5 @@
 const Comment = require('lib/models').Comment
-const privileges = require('lib/models/forum/privileges')
+const privileges = require('lib/privileges/forum')
 const scopes = require('./scopes')
 
 /**
