+++ lib/api-v2/middlewares/forums.js
@@ -1,4 +1,4 @@
-const privileges = require('lib/models/forum/privileges')
+const privileges = require('lib/privileges/forum')
 const config = require('lib/config')
 const api = require('../db-api')
