+++ lib/api/forum/index.js
@@ -6,7 +6,7 @@ var utils = require('lib/utils')
 var restrict = utils.restrict
 var maintenance = utils.maintenance
 var expose = utils.expose
-var privileges = require('lib/models/forum/privileges')
+var privileges = require('lib/privileges/forum')
 var middlewares = require('lib/middlewares/forum-middlewares')
 var Log = require('debug')
