+++ lib/middlewares/forum-middlewares/index.js
@@ -1,6 +1,6 @@
 var config = require('lib/config')
 var api = require('lib/db-api')
-var privileges = require('lib/models/forum/privileges')
+var privileges = require('lib/privileges/forum')
 
 var log = require('debug')('democracyos:forum:middleware')
